Update docker-entrypoint.sh
Removing /bin/bash/ on line 3 allows a docker container to be launched directly from the Docker GUI on Synology, instead of having to do a docker run command via SSH. I tested it by doing a container build with the only change being this line removed. The test image is available on Docker Hub at: thewhitedoll/javinizertest

I believe the #!/bin/bash shebang makes having /bin/bash in the script again redundant.

I'm not entirely sure why this is an issue with the Docker GUI on Synology, but not the command line.